+++ source/host/client_file_transfer.cc
@@ -340,7 +340,7 @@ void ClientFileTransfer::onIpcDisconnected()
 }
 
 //--------------------------------------------------------------------------------------------------
-void ClientFileTransfer::onIpcMessageReceived(quint8 channel_id, const QByteArray& buffer)
+void ClientFileTransfer::onIpcMessageReceived(quint32 channel_id, const QByteArray& buffer)
 {
     if (channel_id != proto::internal::CHANNEL_ID_SESSION)
     {
